+++ app/helpers/application_helper.rb
@@ -74,4 +74,20 @@ module ApplicationHelper
     r, g, b = color.scan(/[a-fA-F0-9]{2}/).map { |color| color.to_i(16) }
     "rgba(#{r}, #{g}, #{b}, #{alpha})"
   end
+
+  def you_tube_link_to(video_id)
+    link_to video_id, "https://www.youtube.com/watch?v=#{video_id}", target: "_blank" if video_id.present?
+  end
+
+  def you_tube_embed(video_id)
+    return if video_id.blank?
+
+    source = "https://www.youtube.com/embed/#{video_id}?&modestbranding=1&showinfo=0&iv_load_policy=3&rel=0"
+
+    content_tag :iframe, nil,
+      src: source,
+      frameborder: 0,
+      allowfullscreen: 1,
+      class: "you-tube-iframe"
+  end
 end