+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170130184418) do
+ActiveRecord::Schema.define(version: 20170314123423) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -354,6 +354,7 @@ ActiveRecord::Schema.define(version: 20170130184418) do
     t.string   "call_to_action",      null: false
     t.integer  "signatures_required", null: false
     t.text     "presentation",        null: false
+    t.string   "video_id"
   end
 
   add_index "petition_plugin_details", ["deleted_at"], name: "index_petition_plugin_details_on_deleted_at", using: :btree